requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
langgraph>=1.1.0,<2.0.0
pydantic>=2.8.0,<3.0.0
python-multipart>=0.0.9,<1.0.0
pillow>=12.0.0,<13.0.0
numpy>=2.0.0,<3.0.0
joblib>=1.4.0,<2.0.0
scikit-learn>=1.5.0,<2.0.0
torch>=2.5.0,<3.0.0
torchvision>=0.20.0,<1.0.0
